+++ src/CompositeData.ts
@@ -22,6 +22,8 @@ export type CompositeProductTimes={
  dwdLightning:ProductTime[];
  dwdRadar?:ProductTime[];
  dwdRadarLayer?:string;
+ dwdRadarLatestOnly?:boolean;
+ mtgLightningLatestOnly?:boolean;
  satelliteDayProduct?:CompositeProduct;
  satelliteIrProduct?:CompositeProduct;
  satellitePrecipProduct?:CompositeProduct;